# openapi.yaml
openapi: 3.1.0
info:
  title: Payment Service
  description: 'Production-ready Digital Wallet payment API. Provides a complete payment processing API including checkouts,
    refunds, customer management, and receipt generation.


    ## Authentication

    All API endpoints (except health checks and auth) require an `X-API-Key` header for authentication. Click the **Authorize**
    button (🔒) above to enter your API key, and it will be included in all requests.


    - **Admin key**: Set via `ADMIN_API_KEY` env var — used to manage tenant API keys

    - **Tenant keys**: Created via `POST /api/v1/admin/api-keys` — used for payment operations


    ## Idempotency

    POST endpoints support the `Idempotency-Key` header to prevent duplicate operations.


    ## Rate Limiting

    API requests are rate-limited per API key. Each tenant key can have custom rate limits. Check the `X-RateLimit-*` response
    headers for current limits.'
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Local development server
paths:
...
  /api/v1/payments:
    post:
      tags:
      - Payments
      summary: Create a payment
      description: Create a new payment via the Digital Wallet system. Supports idempotency via the `Idempotency-Key` header.
      operationId: create_payment_api_v1_payments_post
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Payments
      summary: List payments
      description: Retrieve a paginated list of payments with optional filters.
      operationId: list_payments_api_v1_payments_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 20
          title: Limit
        description: Items per page
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of items to skip
          default: 0
          title: Offset
        description: Number of items to skip
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by status
          title: Status
        description: Filter by status
      - name: customer_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Filter by customer ID
          title: Customer Id
        description: Filter by customer ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/payments/{payment_id}:
    get:
      tags:
      - Payments
      summary: Get payment details
      description: Retrieve details of a specific payment by its ID.
      operationId: get_payment_api_v1_payments__payment_id__get
      parameters:
      - name: payment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Payment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Payments
      summary: Cancel or refund a payment
      description: 'Cancel or refund a payment depending on its current status:


        - **Pending / processing / requires_action** → cancels the payment

        - **Succeeded / partially_refunded** → automatically issues a full refund


        The payment status is updated accordingly (`canceled` or `refunded`).'
      operationId: cancel_or_refund_payment_api_v1_payments__payment_id__delete
      parameters:
      - name: payment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Payment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/payments/{payment_id}/confirm:
    put:
      tags:
      - Payments
      summary: Confirm a payment
      description: Confirm/capture a pending payment intent. Updates the existing payment status to 'succeeded' (or 'requires_action'
        if further steps are needed).
      operationId: confirm_payment_api_v1_payments__payment_id__confirm_put
      parameters:
      - name: payment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Payment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/payments/{payment_id}/receipt:
    get:
      tags:
      - Payments
      summary: Download payment receipt
      description: Generate and download a PDF receipt for a succeeded payment.
      operationId: download_receipt_api_v1_payments__payment_id__receipt_get
      parameters:
      - name: payment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Payment Id
      responses:
        '200':
          description: PDF receipt file
          content:
            application/json:
              schema: {}
            application/pdf: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PaymentCreate:
      properties:
        amount:
          type: integer
          exclusiveMinimum: 0.0
          title: Amount
          description: Amount in smallest currency unit (e.g., cents)
        currency:
          type: string
          maxLength: 3
          minLength: 3
          title: Currency
          description: ISO 4217 currency code
          default: usd
        customer_id:
          type: string
          format: uuid
          title: Customer Id
          description: ID of the customer in our system
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
          description: Payment description
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
          description: Arbitrary key-value metadata
      type: object
      required:
      - amount
      - customer_id
      title: PaymentCreate
      description: Request body to create a new payment.
      examples:
      - amount: 5000
        currency: usd
        customer_id: 550e8400-e29b-41d4-a716-446655440000
        description: Event ticket purchase
        metadata:
          event_id: evt_123
          ticket_type: VIP
    PaymentListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PaymentResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - limit
      - offset
      - has_more
      title: PaymentListResponse
      description: Paginated list of payments.
    PaymentResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        customer_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Customer Id
        amount:
          type: integer
          title: Amount
        currency:
          type: string
          title: Currency
        status:
          type: string
          title: Status
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
        amount_refunded:
          type: integer
          title: Amount Refunded
          default: 0
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - amount
      - currency
      - status
      - created_at
      - updated_at
      title: PaymentResponse
      description: Payment response body.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Enter your API key. Use the **admin key** (from ADMIN_API_KEY env var) for `/api/v1/admin/*` endpoints,
        or a **tenant key** (ps_live_...) for payment/refund/customer operations.
...
security:
- ApiKeyAuth: []